Implement directory tree
This adds functions and a structure for a directory tree, which will
allow for easy recursive directory-ing. It also allows me to sort my
directories when entered from the command line. I still need to make it
not segfault with no options, and I need to print the bad folders, and I
need to make directory adding during the file getting.