Recursion working!
Recursion now properly adds things to the directory tree, by making a
new path to properly ask stat about the file (whether or not it's a
directory), and then being able to properly insert it with it's new
path. I've also added some file-scope variables to tell when I need to
print the newline and the directory name. These will be cleaned up
eventually and put in a place where they make sense, but for now it's
what I'm doing.