Change some file things
file->name is now a pointer to some part of file->path,
to avoid duplication of easily de-duplicated information.
I have also made a new function that sets the name to the
beginning of the path, for when I want the name to not be
truncated (like when printing things gathered from the command
line).