Fix issue with bad path sorting
I had previously tried to solve an issue of path sorting by counting
the number of /'s in each path name while sorting, and sorting more
slashes to the left, because subdirectories should go first. Turns out
that doesn't work with command line, as the things should be
alphabetized first.