fix(makefile): stop tags command from running every time
the ctags command used to happen as a part of the all rule, which meant
that every time I ran make, it regenerated the tags file. Which was
fine, usually it needed to, and it wasn't a problem that it happened.
But 42filechecker checks if your makefile does "something" (i.e.
doesn't say "nothing to be done"), it counts as an error. I'd rather
that not happen, so I made a rule for it. This comes with the added
benefit that now, it stays up to date at the right times. The downside
is that I needed to make the LIB_SRCS makefile variable, which uses
find on the LIB_DIR for all the .c files. This is kinda annoying, but
oh well.