fix(info.c): remove portability, but fix extended attributes
It now does access control lists and extended attributes (cleanly,
without memory leaks), but only for macos. I'm still working on the
linux version, which I might just give up on, because I've gotta
mess about with getxattr (maybe, I'm still not really sure) because
apparently _everything_ on linux has an access control list, so to
determine if it has an _extra_ acl I have to check the extended
attributes. The thing is, I then also need to determine if the
extended attribute that it does have is an acl or not, because files
with extra acl's return something with listxattr, so my function
would automatically put an @ there thinking that it has an extended
attribute, but it could just be an acl. So I could do all that and
it probably wouldn't be that bad, or I could just continue worrying
about it forever and not knowing what to do. I don't know what I'll
do, I just hope I do it soon, whatever it is.