fix(sorting): remove slash check while sorting
I used to check if the differing character while comparing names was a /,
because if it is, I can usually assume that it's a subdirectory (while
going recursively, if the name matches up until then and then it has a /,
then it's definitely a subdirectory, but this isn't always true on the
command line. for instance, when starting with /), but this isn't good
enough. I need to implement subdirectory tracking into my dir struct,
which should be really easy. That's what this branch is for.