feat(-1): add new flag for single column
This adds the new flag to the options string, and also into the enum.
I also update some of the enum comment, since I've taken out the m
maximum size variable. I also gave my enum a type, so I can use it
as a type while dealing with the flags. This is nice because now, i
don't think I have to worry about the size or anything.